requirements.txt
streamlit
matplotlib
pymongo
python-telegram-bot
